modifying name to MainPage, adding Header and Footer
